backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
fuzzywuzzy==0.18.0
python-dateutil==2.8.2
supabase==2.15.3
aiohttp>=3.11.18
schedule==1.2.0
firecrawl-py==0.0.20
posthog==5.4.0
redis==4.5.5
asyncio
# ML dependencies temporarily disabled for clean production deployment
# pandas==2.0.3
# numpy==1.25.1
# scikit-learn==1.3.0
# scipy==1.11.1
# statsmodels==0.14.0
